Add possibility to create skills and intents and use them
The whole system relies on the config-settings, where you can specify which skillId matches which skill in the framework and which skill has which intents.

Additionally its possible to specify actions for a intent. Meaning: When called, the request gets redirected to the skill and an intent (f.e. 'intentName', which then loads the 'SomeClassIntent' Intent) and that Intent calls an action based on the given slots (f.e. the action 'doStuff' may only be called when slots with names 'slot1' and 'slot2' are present).